Make Vocab encoding consistent for python 3
When using Counter in python 3, most_common() is
not guaranteed to return the words that have the same
count in the same order if the code is re run.

This can lead to confusing non reproducible results,
so we simply sort the items in most_common to make sure
the indices returned are always the same.