Refactor model embedding to separate Chain class
Implement general Embedder that allows input to be made up of arbitrary
number of id lists. This refactoring is a step towards implementing a
character level encoder, since we need to simply simply instantiate a
different class as the Embedder.

Major changes:

a) model __call__ function signature has changed.

__call__(*inputs, **kwargs)

where we assume kwargs contains heads and labels if we want to
calculate loss for training or eval purposes.

This allows us to train a model without pos tags if we like - however
we need to change the instantiated Embedder we pass to the model.

NOTE: the reason we used **kwargs and not default arguments is to
retain compatibility with python2 (not allowed to have a signature of
(*args, key=value) in python2.

b) Created Embedder class and moved the embedding code outside Dense
class. This is because we want to slowly transition towards having
subword level embeddings.