Refactor - pull Encoder outside of Dense class
We want to run experiments with different encoders, so we pull the
encoder implementation outside the model in order to be able to
pass it as a parameter.

This also helps scale down the number of parameters we pass directly to
the model.

Also refactored dense tests - use pytest fixtures for easy model creation.