Pad sentences with START and END
Sentences are now padded with START in front and END at end. These
tokens are IDs that are preallocated in all vocabularies implemented in
vocab. We also removed addition of the __ROOT__ token in the dep
loader. Instead, we pad each sentence with START, ROOT and END inside
the model. We also perform all subword padding inside the __call__ of
the model. This makes sense because it is a step that is needed in
general, no matter what the input is. However this approach might not be
very efficient, since we pad everything in the training set multiple
times whilst we could have padded it all in the beginning.