Remove part of speech from input
This is a step towards making the parser less dependent on a part of
speech tagger at test time. For use of previous functionality similar
to what is described in the dissertation the 0.0.1 version can be used.

While it would have been possible to make the use of part of speech as
additional input configurable, it adds significant overhead to code
maintenance for a feature that doesn't really make sense at test time.
The dependency parser is expected to work with raw text input and no
additional information. The plan is to incorporate part of speech
information at training time only by jointly predicting the tags.