Refactor code to functions - make more explicit
Since we are developing a front end application that depends on passing
only text, code that was used for preprocessing the text and encoding
it using the vocabulary was pulled out from larger functions.

Furthermore, named tuples are now used for the vocabulary and input
data to make errors due to ordering of arguments less probable.