Make modular by avoiding premature optimisation
Code used to sort inputs in decreasing order of length to avoid need for
additional computation when feeding and receiving results to and from
LSTM layer. However, this made the code very hard to make modular, as
all the components needed to know what the sorted permutation was.

Also brings the code up to speed for use with more up to date chainer
version.